feature: refactor UserProfileDto package structure and clean up imports
