+++ src/main/java/com/dudoji/spring/dto/user/UserProfileDto.java
@@ -1,4 +1,4 @@
-package com.dudoji.spring.dto;
+package com.dudoji.spring.dto.user;
 
 import com.dudoji.spring.models.domain.User;
